Adding serverproviders

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -7,11 +7,20 @@ use App\Http\Resources\AttendeeResource;
 use Illuminate\Http\Request;
 use App\Models\Event;
 use App\Models\Attendee;
+use App\Http\Traits\CanLoadRelationships;
+
 class AttendeeController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+
+    use CanLoadRelationships;
+
+    private array $relationships = ['user'];
+
+        public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
+
     public function index(Event $event)
     {
         $attendees = $event->attendees()->latest();
@@ -26,9 +35,10 @@ class AttendeeController extends Controller
      */
     public function store(Request $request, Event $event)
     {
-        $attendee = $event->attendees()->create([
+        $this->loadRelationships($attendee = $event->attendees()->create([
             'user_id' => 1,
-        ]);
+        ]));
+
 
         return new AttendeeResource($attendee);
     }
@@ -38,7 +48,8 @@ class AttendeeController extends Controller
      */
     public function show(Event $event, Attendee $attendee)
     {
-        return new AttendeeResource($attendee);
+
+        return new AttendeeResource($this->loadRelationships($attendee));
     }
 
     /**
@@ -54,8 +65,9 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
-         $attendee->delete();
+        $this->authorize('delete-attendee', [$event, $attendee]);
+        $attendee->delete();
 
-         return response(status: 204);
+        return response(status: 204);
     }
 }
